Fix some warnings for printf arguments.
git-svn-id: https://src.develer.com/svnoss/bertos/trunk@1753 38d2e660-2303-0410-9eaa-f027e97ec537

+++ bertos/drv/flash25.c
@@ -325,8 +325,7 @@ void flash25_sectorErase(Flash25 *fd, Flash25Sector sector)
 	 */
 	flash25_waitReady(fd);
 
-	DB(kprintf("Erased sector [%ld] in %d ms\n", sector, ticks_to_ms(timer_clock() - start_time)));
-
+	DB(kprintf("Erased sector [%ld] in %ld ms\n", (unsigned long)sector, (unsigned long)ticks_to_ms(timer_clock() - start_time)));
 }
 
 /**
